+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Keeps track of notable changes. Please remember to add new behaviours to the
 Unreleased section to make new releases easy.
 
 ## [Unreleased]
+
+### Added
+
+- Added railties to automatically initialize and configure Warren in rails apps.
+  You can remove the `config/initializers/warren.rb` and the `config.warren = config_for(:warren)`
+  line from `config/application.rb`
+
 ## Initial release
 
 - Import of lib/warren from sequencescape